Moved markup files to their own package
