+++ markup/markup_asciidoc.go
@@ -1,4 +1,4 @@
-package main
+package markup
 
 import (
 	"fmt"
